fix: 알림 항목 변경, 영양제 -> 비료 / 분갈이 -> 건강확인
