--- src/main/kotlin/dnd11th/blooming/api/dto/AlarmModifyRequest.kt
@@ -1,22 +1,0 @@
-package dnd11th.blooming.api.dto
-
-import dnd11th.blooming.domain.entity.Alarm
-
-data class AlarmModifyRequest(
-    val waterAlarm: Boolean,
-    val waterPeriod: Int?,
-    val nutrientsAlarm: Boolean,
-    val nutrientsPeriod: Int?,
-    val repotAlarm: Boolean,
-    val repotPeriod: Int?,
-) {
-    fun toAlarm(): Alarm =
-        Alarm(
-            waterAlarm = waterAlarm,
-            waterPeriod = waterPeriod,
-            nutrientsAlarm = nutrientsAlarm,
-            nutrientsPeriod = nutrientsPeriod,
-            repotAlarm = repotAlarm,
-            repotPeriod = repotPeriod,
-        )
-}
