+++ src/main/kotlin/dnd11th/blooming/api/dto/myplant/MyPlantSaveRequest.kt
@@ -1,4 +1,4 @@
-package dnd11th.blooming.api.dto
+package dnd11th.blooming.api.dto.myplant
 
 import dnd11th.blooming.domain.entity.Alarm
 import dnd11th.blooming.domain.entity.MyPlant
@@ -11,10 +11,9 @@ data class MyPlantSaveRequest(
     val lastWateredDate: LocalDate,
     val waterAlarm: Boolean,
     val waterPeriod: Int?,
-    val nutrientsAlarm: Boolean,
-    val nutrientsPeriod: Int?,
-    val repotAlarm: Boolean,
-    val repotPeriod: Int?,
+    val fertilizerAlarm: Boolean,
+    val fertilizerPeriod: Int?,
+    val healthCheckAlarm: Boolean,
 ) {
     fun toMyPlant(): MyPlant =
         MyPlant(
@@ -26,10 +25,9 @@ data class MyPlantSaveRequest(
                 Alarm(
                     waterAlarm = waterAlarm,
                     waterPeriod = waterPeriod,
-                    nutrientsAlarm = nutrientsAlarm,
-                    nutrientsPeriod = nutrientsPeriod,
-                    repotAlarm = repotAlarm,
-                    repotPeriod = repotPeriod,
+                    fertilizerAlarm = fertilizerAlarm,
+                    fertilizerPeriod = fertilizerPeriod,
+                    healthCheckAlarm = healthCheckAlarm,
                 ),
         )
 }
